FastAPI/02/pydantic/pyproject.toml
[tool.poetry]
name = "data_classes"
version = "0.1.0"
description = "Esplorando le data classes, con pydantic"
authors = ["Burlesco70"]

[tool.poetry.dependencies]
python = "^3.8"
python-dateutil = "^2.8.1"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
